updating remove orderItem function by passing history object from useHistory hook
