+++ client/store/orders.js
@@ -1,10 +1,8 @@
 import Axios from 'axios';
 
-
-const GET_ORDERS = "GET_ORDERS";
+const GET_ORDERS = 'GET_ORDERS';
 
 const COMPLETE_ORDER = 'COMPLETE_ORDER';
-const DELETE_ORDERITEM = 'DELETE_ORDERITEM';
 
 export const getOrders = (orders) => {
   return {
@@ -17,12 +15,6 @@ export const completeOrder = (order) => {
   return {
     type: COMPLETE_ORDER,
     order,
-  };
-};
-export const deleteOrderItem = (orderItem) => {
-  return {
-    type: DELETE_ORDERITEM,
-    orderItem,
   };
 };
 
@@ -57,35 +49,13 @@ export const completeNewOrder = (orderId, history) => {
         });
         dispatch(completeOrder(data));
 
-        history.push("/orders");
+        history.push('/orders');
       }
     } catch (err) {
       console.log(err);
     }
   };
 };
-
-export const fetchDeleteOrderItem = (productId) => {
-  return async (dispatch) => {
-    try {
-      const token = window.localStorage.getItem('token');
-
-      if (token) {
-        const { data } = await Axios.delete('/api/orders', {
-          headers: {
-            authorization: token,
-          },
-          data: productId,
-        });
-
-        dispatch(deleteOrderItem(data));
-      }
-    } catch (err) {
-      console.error(err);
-    }
-  };
-};
-
 
 const intialState = {
   completeOrder: [],
@@ -98,8 +68,6 @@ export default function orderReducer(state = intialState, action) {
     case COMPLETE_ORDER:
       state.completeOrder = [...state.completeOrder, action.order];
       return state;
-    case DELETE_ORDERITEM:
-      return { ...state, newOrder: action.orderItem };
     default:
       return state;
   }
